requirements.txt
click==8.1.2
cycler==0.11.0
fonttools==4.31.2
h5py==3.6.0
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scipy==1.8.0
six==1.16.0
typer==0.4.1
